lcb: Generated predicates for arithmetical field access functions
