+++ vadl/main/vadl/gcb/passes/encodingGeneration/strategies/EncodingPredicateGenerationStrategy.java
@@ -16,17 +16,31 @@
 
 package vadl.gcb.passes.encodingGeneration.strategies;
 
+import java.math.BigInteger;
+import vadl.gcb.passes.GenerateValueRangeImmediatePass;
+import vadl.types.BuiltInTable;
 import vadl.types.Type;
+import vadl.utils.GraphUtils;
+import vadl.utils.SourceLocation;
+import vadl.viam.Constant;
 import vadl.viam.Format;
 import vadl.viam.Function;
+import vadl.viam.Identifier;
 import vadl.viam.Instruction;
 import vadl.viam.Parameter;
 import vadl.viam.graph.Graph;
+import vadl.viam.graph.control.ReturnNode;
+import vadl.viam.graph.control.StartNode;
+import vadl.viam.graph.dependency.ConstantNode;
+import vadl.viam.graph.dependency.FuncParamNode;
 
 /**
  * The implementor of this interface can generate a field access encoding / predicate functions.
  */
 public interface EncodingPredicateGenerationStrategy {
+
+  String PARAM = "param";
+
   /**
    * Check if the strategy can be applied. Returns {@code true} when it is applicable.
    */
@@ -68,5 +82,42 @@ public interface EncodingPredicateGenerationStrategy {
         .parts());
     var predicate = new Function(ident, new Parameter[] {}, Type.bool(), behavior);
     fieldAccess.setPredicate(predicate);
+  }
+
+  /**
+   * Template method for generation a predicate function.
+   */
+  default void generatePredicate(Instruction instruction, Format.FieldAccess fieldAccess) {
+    var trueCase = new ConstantNode(Constant.Value.fromBoolean(true));
+    var falseCase = new ConstantNode(Constant.Value.fromBoolean(false));
+
+    var fieldRef = fieldAccess.fieldRefs().getFirst();
+
+    var paramNode = new FuncParamNode(
+        new Parameter(new Identifier(PARAM, SourceLocation.INVALID_SOURCE_LOCATION),
+            fieldAccess.type()));
+
+    var lowestValue = new ConstantNode(Constant.Value.fromInteger(
+        BigInteger.valueOf(
+            GenerateValueRangeImmediatePass.lowestPossibleValue(fieldRef.type().toBitsType(), true)),
+        Type.signedInt(64)));
+    var highestValue = new ConstantNode(Constant.Value.fromInteger(
+        BigInteger.valueOf(
+            GenerateValueRangeImmediatePass.highestPossibleValue(fieldRef.type().toBitsType(), true)),
+        Type.signedInt(64)));
+
+    var lowestExpr =
+        GraphUtils.binaryOp(BuiltInTable.SGEQ, paramNode, lowestValue);
+    var highestExpr =
+        GraphUtils.binaryOp(BuiltInTable.SLEQ, paramNode, highestValue);
+
+    var conditional =
+        GraphUtils.select(GraphUtils.and(lowestExpr, highestExpr), trueCase, falseCase);
+    var returnNode = new ReturnNode(conditional);
+    var startNode = new StartNode(returnNode);
+    var behavior = new Graph("Generated predicate of " + fieldAccess.simpleName());
+    behavior.addWithInputs(returnNode);
+    behavior.add(startNode);
+    setPredicate(instruction, fieldAccess, behavior);
   }
 }
