+++ vadl/main/vadl/gcb/passes/encodingGeneration/strategies/impl/ArithmeticImmediateStrategyPredicate.java
@@ -24,7 +24,6 @@ import vadl.types.DataType;
 import vadl.viam.Constant;
 import vadl.viam.Format;
 import vadl.viam.Instruction;
-import vadl.viam.PrintableInstruction;
 import vadl.viam.graph.GraphVisitor;
 import vadl.viam.graph.Node;
 import vadl.viam.graph.control.ReturnNode;
@@ -71,7 +70,13 @@ public class ArithmeticImmediateStrategyPredicate implements EncodingPredicateGe
   }
 
   @Override
-  public void generateEncodingAndPredicateFunction(Instruction instruction, Format.FieldAccess fieldAccess) {
+  public void generateEncodingAndPredicateFunction(Instruction instruction,
+                                                   Format.FieldAccess fieldAccess) {
+    generateEncoding(instruction, fieldAccess);
+    generatePredicate(instruction, fieldAccess);
+  }
+
+  private void generateEncoding(Instruction instruction, Format.FieldAccess fieldAccess) {
     var accessFunction = fieldAccess.accessFunction();
     var copy = accessFunction.behavior().copy();
     final var returnNode = copy.getNodes(ReturnNode.class).findFirst().get();
@@ -87,7 +92,7 @@ public class ArithmeticImmediateStrategyPredicate implements EncodingPredicateGe
 
     // After that we need to find the field and add it to the other side.
     var fieldRefs = copy.getNodes(FieldRefNode.class).toList();
-    var fieldRef = fieldRefs.get(0);
+    var fieldRef = fieldRefs.getFirst();
     var fieldRefBits = (BitsType) fieldRef.type();
 
     // Example
